requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.7
defusedxml==0.7.1
distlib==0.3.8
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.61.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==24.2
django-js-asset==2.2.0
django-mptt==0.16.0
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
drf-yasg==1.21.7
filelock==3.13.1
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
macholib==1.16.3
MarkupSafe==2.1.5
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.0
pefile==2023.2.7
pillow==10.3.0
pip-system-certs==4.0
platformdirs==4.1.0
pycparser==2.22
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.4
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.4
pywin32-ctypes==0.2.2
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
simplejwt==2.0.1
six==1.16.0
sqlparse==0.5.0
typing==3.7.4.3
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.0
wrapt==1.16.0
